One dashboard per wiki, more actions, more datasources, more graphs
This is a big reorg of our limn stats. Instead of putting all wikis
on the same dashboard, we have display global stats on /mmv
and display per wiki stats on /mmv_wiki dashboard.

This also allows us to split up the performance graphs per call
which leaves room to display the percentile options for each.